requirement.txt
beautifulsoup4==4.9.1
certifi==2020.6.20
chardet==3.0.4
colorama==0.4.4
configparser==5.0.2
crayons==0.4.0
cssselect==1.1.0
Django==2.2.2
django-cors-headers==3.5.0
djangorestframework==3.9.2
fake-useragent==0.1.11
future==0.18.2
google==3.0.0
Google-Search-API==1.1.14
idna==2.10
lxml==4.5.2
multidict==6.0.2
parsel==1.6.0
pytz==2020.1
PyYAML==6.0
requests==2.25.1
selenium==2.53.6
selenium-wire==2.1.1
six==1.15.0
soupsieve==2.0.1
sqlparse==0.3.1
Unidecode==1.3.4
urllib3==1.25.10
vcrpy==4.1.1
w3lib==1.22.0
webdriver-manager==3.3.0
wrapt==1.14.1
yarl==1.7.2
